project structure: changed scraping service directory name
