+++ infra/scraper/src/scraper/exceptions.py
