+++ infra/scraper/src/scraper/models.py
